Refactor physac.go

+++ main.go
@@ -66,8 +66,8 @@ func (g *Game) Draw(screen *ebiten.Image) {
 			vertexB := GetPhysicsShapeVertex(body, jj)
 
 			ebitenutil.DrawLine(screen,
-				float64(vertexA.X), float64(vertexA.Y),
-				float64(vertexB.X), float64(vertexB.Y),
+				vertexA.X, vertexA.Y,
+				vertexB.X, vertexB.Y,
 				color.RGBA{0, 255, 0, 255},
 			)
 		}
@@ -85,23 +85,28 @@ func main() {
 	ebiten.SetWindowTitle("Physac golang demo")
 
 	// Create floor rectangle physics body
-	floor := CreatePhysicsBodyRectangle(Vector2{screenWidth / 2, screenHeight * 0.975}, screenWidth*1.3, screenHeight/5, 10)
+	floor := CreatePhysicsBodyRectangle(Vector2{screenWidth / 2, screenHeight * 0.975},
+		screenWidth*1.3, screenHeight/5, 10)
 	SetPhysicsBodyRotation(floor, (math.Pi/180)*3)
 	floor.Enabled = false // Disable body state to convert it to static (no dynamics, but collisions)
 	floor.Restitution = 1
 
-	floor2 := CreatePhysicsBodyRectangle(Vector2{screenWidth / 2, screenHeight * 0.975}, screenWidth*1.3, screenHeight/5, 10)
+	floor2 := CreatePhysicsBodyRectangle(Vector2{screenWidth / 2, screenHeight * 0.975},
+		screenWidth*1.3, screenHeight/5, 10)
 	SetPhysicsBodyRotation(floor2, (math.Pi/180)*-3)
 	floor2.Enabled = false // Disable body state to convert it to static (no dynamics, but collisions)
 	floor2.Restitution = 1
 
-	circle := CreatePhysicsBodyCircle(Vector2{screenWidth * 0.10, screenHeight * 0.1}, screenWidth*0.02, 10)
+	circle := CreatePhysicsBodyCircle(Vector2{screenWidth * 0.10, screenHeight * 0.1},
+		screenWidth*0.02, 10)
 	circle.Restitution = 0.5
 
-	circle2 := CreatePhysicsBodyCircle(Vector2{screenWidth * 0.90, screenHeight * 0.2}, screenWidth*0.03, 10)
+	circle2 := CreatePhysicsBodyCircle(Vector2{screenWidth * 0.90, screenHeight * 0.2},
+		screenWidth*0.03, 10)
 	circle2.Restitution = 0.3
 
-	circle3 := CreatePhysicsBodyCircle(Vector2{screenWidth * 0.50, screenHeight * 0.05}, screenWidth*0.015, 10)
+	circle3 := CreatePhysicsBodyCircle(Vector2{screenWidth * 0.50, screenHeight * 0.05},
+		screenWidth*0.015, 10)
 	circle3.Restitution = 1
 
 	SetPhysicsGravity(0, 9.81/10)
